Implement our Token and ResolvedToken classes
Our Token class is a "raw token" and is extremely basic.
It basically exists to define the name of the token for the parser and for a method `getValue` that will return a source value out of the data dict.

The ResolvedToken is a bit more interesting, it contains a single "raw token" and zero or more operator instances.
This is not a public class and should not be subclassed. It gets automatically constructed for the tokens in the path and executed on by the parser class.