Implement concrete Template class
The entrypoint to this class is the `from_template_str` classmethod.
It requires your template_str as well as your TokenManager and OperatorManager instances.
Because of this, you probably don't want to instantiate this class except from the Parser class.

We first resolve your template_str to a set of tokens using our PEG parser.
After that, we convert the "raw token" and Operators into ResolvedToken objects, store them on the Template object and store the source template_str.

The public `resolve` method takes data and will sequentially resolve the template_str using the data provided.
By this point, because the Template object has been instantiated, you should be safe in assuming that your template_str is valid and contains valid Operators and Tokens.
However, if your Operators pass bad data through to each other, you may get issues when executing the ResolvedToken objects.